backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
jaydebeapi>=1.2.3
JPype1>=1.5.0
pydantic>=2.5.0
